Move Produce, Rideable, & Rider into interfaces package
