+++ src/main/java/com/zipcodewilmington/froilansfarm/Aircraft.java
@@ -1,6 +1,7 @@
 package com.zipcodewilmington.froilansfarm;
 
 import com.zipcodewilmington.froilansfarm.interfaces.NoiseMaker;
+import com.zipcodewilmington.froilansfarm.interfaces.Rideable;
 
 public abstract class Aircraft<F extends Vehicle> implements Rideable, NoiseMaker {
