+++ src/main/java/com/zipcodewilmington/froilansfarm/Chicken.java
@@ -1,5 +1,7 @@
 package com.zipcodewilmington.froilansfarm;
 
+import com.zipcodewilmington.froilansfarm.interfaces.Produce;
+
 public class Chicken extends Animal implements Produce {
 
     int legs;
