+++ src/main/java/com/zipcodewilmington/froilansfarm/Horse.java
@@ -1,6 +1,8 @@
 package com.zipcodewilmington.froilansfarm;
 
-public class Horse extends Animal implements Rideable{
+import com.zipcodewilmington.froilansfarm.interfaces.Rideable;
+
+public class Horse extends Animal implements Rideable {
 
     int legs;
