--- src/main/java/com/zipcodewilmington/froilansfarm/Rideable.java
@@ -1,8 +1,0 @@
-package com.zipcodewilmington.froilansfarm;
-
-public interface Rideable {
-
-
-
-
-}
