+++ src/main/java/com/zipcodewilmington/froilansfarm/Vehicle.java
@@ -1,6 +1,7 @@
 package com.zipcodewilmington.froilansfarm;
 
 import com.zipcodewilmington.froilansfarm.interfaces.NoiseMaker;
+import com.zipcodewilmington.froilansfarm.interfaces.Rideable;
 
 public abstract class Vehicle implements Rideable, NoiseMaker {
