+++ src/main/java/com/zipcodewilmington/froilansfarm/interfaces/Rideable.java
@@ -1,0 +1,8 @@
+package com.zipcodewilmington.froilansfarm.interfaces;
+
+public interface Rideable {
+
+
+
+
+}
