+++ src/main/java/com/zipcodewilmington/froilansfarm/interfaces/Rider.java
@@ -1,4 +1,4 @@
-package com.zipcodewilmington.froilansfarm;
+package com.zipcodewilmington.froilansfarm.interfaces;
 
 public interface Rider {
